refactor: split self-defined alphas to `alpha.py`
